refactor!: split actions across multiple files
Split up the "actions.lua" file and put the corresponding functions into
separate files with a more narrow scope. Also extract some of functions
from "drex.lua" which deal with element handling (expanding, collapsing,
focus) into a separate "actions/elements.lua" file

The goal is to make the project structure simpler, easier to maintain
and especially easier to expand
