+++ lua/drex/config/init.lua
@@ -38,57 +38,57 @@ local defaults = {
             -- always use visual mode linewise for better visibility
             ['v'] = 'V',
             -- open/close directories
-            ['l'] = '<cmd>lua require("drex").expand_element()<CR>',
-            ['h'] = '<cmd>lua require("drex").collapse_directory()<CR>',
-            ['<right>'] = '<cmd>lua require("drex").expand_element()<CR>',
-            ['<left>']  = '<cmd>lua require("drex").collapse_directory()<CR>',
-            ['<2-LeftMouse>'] = '<LeftMouse><cmd>lua require("drex").expand_element()<CR>',
-            ['<RightMouse>']  = '<LeftMouse><cmd>lua require("drex").collapse_directory()<CR>',
+            ['l'] = '<cmd>lua require("drex.actions.elements").expand_element()<CR>',
+            ['h'] = '<cmd>lua require("drex.actions.elements").collapse_directory()<CR>',
+            ['<right>'] = '<cmd>lua require("drex.actions.elements").expand_element()<CR>',
+            ['<left>']  = '<cmd>lua require("drex.actions.elements").collapse_directory()<CR>',
+            ['<2-LeftMouse>'] = '<LeftMouse><cmd>lua require("drex.actions.elements").expand_element()<CR>',
+            ['<RightMouse>']  = '<LeftMouse><cmd>lua require("drex.actions.elements").collapse_directory()<CR>',
             -- open files in separate windows/tabs
-            ['<C-v>'] = '<cmd>lua require("drex").open_file("vs")<CR>',
-            ['<C-x>'] = '<cmd>lua require("drex").open_file("sp")<CR>',
-            ['<C-t>'] = '<cmd>lua require("drex").open_file("tabnew", true)<CR>',
+            ['<C-v>'] = '<cmd>lua require("drex.actions.elements").open_file("vs")<CR>',
+            ['<C-x>'] = '<cmd>lua require("drex.actions.elements").open_file("sp")<CR>',
+            ['<C-t>'] = '<cmd>lua require("drex.actions.elements").open_file("tabnew", true)<CR>',
             -- switch root directory
-            ['<C-l>'] = '<cmd>lua require("drex").open_directory()<CR>',
-            ['<C-h>'] = '<cmd>lua require("drex").open_parent_directory()<CR>',
+            ['<C-l>'] = '<cmd>lua require("drex.actions.elements").open_directory()<CR>',
+            ['<C-h>'] = '<cmd>lua require("drex.actions.elements").open_parent_directory()<CR>',
             -- manual reload
             ['<F5>'] = '<cmd>lua require("drex").reload_directory()<CR>',
             -- jump around elements
-            ['gj'] = '<cmd>lua require("drex.jump").jump_to_next_sibling()<CR>',
-            ['gk'] = '<cmd>lua require("drex.jump").jump_to_prev_sibling()<CR>',
-            ['gh'] = '<cmd>lua require("drex.jump").jump_to_parent()<CR>',
+            ['gj'] = '<cmd>lua require("drex.actions.jump").jump_to_next_sibling()<CR>',
+            ['gk'] = '<cmd>lua require("drex.actions.jump").jump_to_prev_sibling()<CR>',
+            ['gh'] = '<cmd>lua require("drex.actions.jump").jump_to_parent()<CR>',
             -- file actions
-            ['s'] = '<cmd>lua require("drex.actions").stats()<CR>',
-            ['a'] = '<cmd>lua require("drex.actions").create()<CR>',
-            ['d'] = '<cmd>lua require("drex.actions").delete("line")<CR>',
-            ['D'] = '<cmd>lua require("drex.actions").delete("clipboard")<CR>',
-            ['p'] = '<cmd>lua require("drex.actions").copy_and_paste()<CR>',
-            ['P'] = '<cmd>lua require("drex.actions").cut_and_move()<CR>',
-            ['r'] = '<cmd>lua require("drex.actions").rename()<CR>',
-            ['R'] = '<cmd>lua require("drex.actions").multi_rename("clipboard")<CR>',
+            ['s'] = '<cmd>lua require("drex.actions.stats").stats()<CR>',
+            ['a'] = '<cmd>lua require("drex.actions.files").create()<CR>',
+            ['d'] = '<cmd>lua require("drex.actions.files").delete("line")<CR>',
+            ['D'] = '<cmd>lua require("drex.actions.files").delete("clipboard")<CR>',
+            ['p'] = '<cmd>lua require("drex.actions.files").copy_and_paste()<CR>',
+            ['P'] = '<cmd>lua require("drex.actions.files").cut_and_move()<CR>',
+            ['r'] = '<cmd>lua require("drex.actions.files").rename()<CR>',
+            ['R'] = '<cmd>lua require("drex.actions.files").multi_rename("clipboard")<CR>',
             -- add/remove elements from clipboard
             ['M'] = '<cmd>DrexMark<CR>',
             ['u'] = '<cmd>DrexUnmark<CR>',
             ['m'] = '<cmd>DrexToggle<CR>',
-            ['cc'] = '<cmd>lua require("drex.actions").clear_clipboard()<CR>',
-            ['cs'] = '<cmd>lua require("drex.actions").open_clipboard_window()<CR>',
+            ['cc'] = '<cmd>lua require("drex.actions.clipboard").clear_clipboard()<CR>',
+            ['cs'] = '<cmd>lua require("drex.actions.clipboard").open_clipboard_window()<CR>',
             -- string copy utilities
-            ['y'] = '<cmd>lua require("drex.actions").copy_element_name()<CR>',
-            ['Y'] = '<cmd>lua require("drex.actions").copy_element_relative_path()<CR>',
-            ['<C-y>'] = '<cmd>lua require("drex.actions").copy_element_absolute_path()<CR>',
+            ['y'] = '<cmd>lua require("drex.actions.text").copy_name()<CR>',
+            ['Y'] = '<cmd>lua require("drex.actions.text").copy_relative_path()<CR>',
+            ['<C-y>'] = '<cmd>lua require("drex.actions.text").copy_absolute_path()<CR>',
         },
         ['v'] = {
             -- file actions
-            ['d'] = ':lua require("drex.actions").delete("visual")<CR>',
-            ['r'] = ':lua require("drex.actions").multi_rename("visual")<CR>',
+            ['d'] = ':lua require("drex.actions.files").delete("visual")<CR>',
+            ['r'] = ':lua require("drex.actions.files").multi_rename("visual")<CR>',
             -- add/remove elements from clipboard
             ['M'] = ':DrexMark<CR>',
             ['u'] = ':DrexUnmark<CR>',
             ['m'] = ':DrexToggle<CR>',
             -- string copy utilities
-            ['y'] = ':lua require("drex.actions").copy_element_name(true)<CR>',
-            ['Y'] = ':lua require("drex.actions").copy_element_relative_path(true)<CR>',
-            ['<C-y>'] = ':lua require("drex.actions").copy_element_absolute_path(true)<CR>',
+            ['y'] = ':lua require("drex.actions.text").copy_name(true)<CR>',
+            ['Y'] = ':lua require("drex.actions.text").copy_relative_path(true)<CR>',
+            ['<C-y>'] = ':lua require("drex.actions.text").copy_absolute_path(true)<CR>',
         }
     },
     on_enter = nil,
